uploader/requirements.txt
Django==4.1.10
djangorestframework==3.13.1
mysql-connector-python==8.2.0
pillow==10.4.0
prometheus-client==0.19.0
python-decouple==3.8
mysqlclient==2.2.0